♻️ Migrate the prompt provider which has versioning feature into another module *provider* to converge the relative logic.
